requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
avalanche-lib==0.6.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.0
click==8.2.1
cmeel==0.57.3
contourpy==1.3.2
cryptography==44.0.3
cvxopt==1.3.2
cycler==0.12.1
daqp==0.7.1
datasets==3.1.0
dill==0.3.8
docker-pycreds==0.4.0
ecos==2.0.14
filelock==3.18.0
flwr==1.18.0
flwr-datasets==0.5.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2024.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
grpcio==1.71.0
hf-xet==1.1.2
highspy==1.10.0
huggingface-hub==0.32.3
hydra-core==1.3.2
idna==3.10
iterators==0.0.2
jax==0.6.1
jaxlib==0.6.1
jaxopt==0.8.5
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opt_einsum==3.4.0
osqp==1.0.4
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
piqp==0.5.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==4.25.8
proxsuite==0.7.2
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorchcv==0.0.73
pytz==2025.2
PyYAML==6.0.2
qpalm==1.2.5
qpax==0.0.9
qpsolvers==4.7.0
quadprog==0.1.13
ray==2.46.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
scs==3.2.7.post2
seaborn==0.13.2
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==78.1.1
shellingham==1.5.4
sip_python==0.0.2
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tomli==2.2.1
tomli_w==1.2.0
torch==2.1.0
torchmetrics==1.7.2
torchvision==0.16.0
tqdm==4.67.1
triton==2.1.0
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
